Completed Issue 11 user can login. -Aaron Woo

+++ server/index.js
@@ -6,6 +6,8 @@ const pg = require('pg');
 const ClientError = require('./client-error');
 const uploadsMiddleware = require('./uploads-middleware');
 const argon2 = require('argon2');
+const jwt = require('jsonwebtoken');
+const authorizationMiddleware = require('./authorization-middleware');
 
 const db = new pg.Pool({
   connectionString: process.env.DATABASE_URL,
@@ -41,14 +43,49 @@ app.post('/api/auth/sign-up', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.post('/api/inventory/1', uploadsMiddleware, (req, res, next) => {
+app.post('/api/auth/sign-in', (req, res, next) => {
+  const { username, password: userPassword } = req.body;
+  if (!username || !userPassword) {
+    throw new ClientError(401, 'invalid login');
+  }
+  const sql = `
+    select "userId",
+           "password"
+      from "users"
+     where "username" = $1
+  `;
+  const params = [username];
+  db.query(sql, params)
+    .then(result => {
+      const [user] = result.rows;
+      if (!user) {
+        throw new ClientError(401, 'invalid login');
+      }
+      const { userId, password } = user;
+      return argon2
+        .verify(password, userPassword)
+        .then(isMatching => {
+          if (!isMatching) {
+            throw new ClientError(401, 'invalid login');
+          }
+          const payload = { userId, username };
+          const token = jwt.sign(payload, process.env.TOKEN_SECRET);
+          res.json({ token, user: payload });
+        });
+    })
+    .catch(err => next(err));
+});
+
+app.use(authorizationMiddleware);
+
+app.post('/api/inventory', uploadsMiddleware, (req, res, next) => {
   const {
     articleTypeId, primaryColor, secondaryColor, colorCategoryId, secondaryColorCategoryId
   } = req.body;
   if (!articleTypeId || !primaryColor || !colorCategoryId) {
     throw new ClientError(401, 'Invalid article, articleTypeId, primaryColor, and colorCategoryId are required.');
   }
-
+  const userId = req.user.userId;
   const imgUrl = `/images/${req.file.filename}`;
 
   const sql = `
@@ -57,7 +94,7 @@ app.post('/api/inventory/1', uploadsMiddleware, (req, res, next) => {
                 values ($1, $2, $3, $4, $5, $6, $7)
                 returning *
   `;
-  const params = [1, imgUrl, articleTypeId, primaryColor, secondaryColor, colorCategoryId, secondaryColorCategoryId];
+  const params = [userId, imgUrl, articleTypeId, primaryColor, secondaryColor, colorCategoryId, secondaryColorCategoryId];
   db.query(sql, params)
     .then(result => {
       const [articles] = result.rows;
@@ -66,14 +103,16 @@ app.post('/api/inventory/1', uploadsMiddleware, (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.delete('/api/inventory/1/:articleId', (req, res, next) => {
+app.delete('/api/inventory/:articleId', (req, res, next) => {
   const articleId = req.params.articleId;
+  const userId = req.user.userId;
   const sql = `
     delete from "articles"
       where "articleId" = $1
+      AND "userId" = $2
       returning *;
   `;
-  const params = [articleId];
+  const params = [articleId, userId];
   db.query(sql, params)
     .then(result => {
       const [article] = result.rows;
@@ -85,9 +124,9 @@ app.delete('/api/inventory/1/:articleId', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.post('/api/outfits/1', (req, res, next) => {
+app.post('/api/outfits', (req, res, next) => {
   const { topArticleId, bottomArticleId, shoesArticleId } = req.body;
-  const userId = 1;
+  const userId = req.user.userId;
   if (!topArticleId || !bottomArticleId || !shoesArticleId) {
     throw new ClientError(401, 'Invalid outfit, requires top, bottom, and shoes');
   }
@@ -107,16 +146,18 @@ app.post('/api/outfits/1', (req, res, next) => {
 
 app.use(staticMiddleware);
 
-app.get('/api/inventory/1', (req, res, next) => {
+app.get('/api/inventory/:userId', (req, res, next) => {
+  const userId = req.params.userId;
   const sql = `
     select "articleId",
           "imgUrl",
           "primaryColor",
           "secondaryColor"
       from "articles"
-      where "userId" = 1
+      where "userId" = $1
   `;
-  db.query(sql)
+  const params = [userId];
+  db.query(sql, params)
     .then(result => {
       if (result.rows.length === 0) {
         res.json([]);
@@ -127,8 +168,9 @@ app.get('/api/inventory/1', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.get('/api/inventory/1/:articleType', (req, res, next) => {
+app.get('/api/inventory/:userId/:articleType', (req, res, next) => {
   const articleType = req.params.articleType;
+  const userId = req.params.userId;
   const articleTypeIds = {
     tops: 1,
     bottoms: 2,
@@ -141,10 +183,10 @@ app.get('/api/inventory/1/:articleType', (req, res, next) => {
             "primaryColor",
             "secondaryColor"
         from "articles"
-        where "userId" = 1
-        AND "articleTypeId" = $1
+        where "userId" = $1
+        AND "articleTypeId" = $2
   `;
-  const params = [articleTypeId];
+  const params = [userId, articleTypeId];
   db.query(sql, params)
     .then(result => {
       if (result.rows.length === 0) {
@@ -156,8 +198,9 @@ app.get('/api/inventory/1/:articleType', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.get('/api/inventory/1/:articleType/:color', (req, res, next) => {
+app.get('/api/inventory/:userId/:articleType/:color', (req, res, next) => {
   const articleType = req.params.articleType;
+  const userId = req.params.userId;
   const color = req.params.color;
   const articleTypeIds = {
     tops: 1,
@@ -186,11 +229,11 @@ app.get('/api/inventory/1/:articleType/:color', (req, res, next) => {
         "secondaryColor",
         "articleTypeId"
     from "articles"
-    where "userId" = 1
-    AND "articleTypeId" = $1
-    AND ("colorCategoryId" = $2 OR "secondaryColorCategoryId" = $2);
+    where "userId" = $1
+    AND "articleTypeId" = $2
+    AND ("colorCategoryId" = $3 OR "secondaryColorCategoryId" = $3);
   `;
-  const params = [articleTypeId, colorId];
+  const params = [userId, articleTypeId, colorId];
   db.query(sql, params)
     .then(result => {
       if (result.rows.length === 0) {
@@ -210,7 +253,8 @@ app.get('/api/inventory/1/:articleType/:color', (req, res, next) => {
     .catch(err => next(err));
 });
 
-app.get('/api/outfits/1', (req, res, next) => {
+app.get('/api/outfits/:userId', (req, res, next) => {
+  const userId = req.params.userId;
   const sql = `
       select "o"."outfitId",
         "a"."imgUrl",
@@ -219,12 +263,14 @@ app.get('/api/outfits/1', (req, res, next) => {
         "a"."secondaryColor"
     from "articles" as "a"
     join "outfits" as "o" using ("userId")
-    where "o"."topArticleId" = "a"."articleId"
+    where "o"."userId" = $1
+    AND ("o"."topArticleId" = "a"."articleId"
     OR "o"."bottomArticleId" = "a"."articleId"
-    OR "o"."shoesArticleId" = "a"."articleId"
+    OR "o"."shoesArticleId" = "a"."articleId")
     order by "o"."outfitId"
   `;
-  db.query(sql)
+  const params = [userId];
+  db.query(sql, params)
     .then(result => {
       if (result.rows.length === 0) {
         res.json([]);
@@ -234,13 +280,14 @@ app.get('/api/outfits/1', (req, res, next) => {
       const outfitsArr = [];
       let currOutfit = [];
       for (let i = 0; i < articlesData.length; i++) {
+        currOutfit.push(articlesData[i]);
         if (currOutfit.length === 3) {
           outfitsArr.push(currOutfit);
           currOutfit = [];
         }
-        currOutfit.push(articlesData[i]);
       }
       res.json(outfitsArr);
+
     })
     .catch(err => next(err));
 });
